requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.7
pypdf==4.3.1
openpyxl==3.1.5
pytest==8.4.1
httpx==0.28.1
Pillow==10.4.0
pytesseract==0.3.13
edge-tts==6.1.13
